+++ src/main/java/kr/hhplus/be/server/point/api/PointChargeRequest.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.point;
+package kr.hhplus.be.server.point.api;
 
 import lombok.Getter;
 import lombok.Setter;
